GRT-68 [MOVE] Tag 컴포넌트 common 폴더 아래로 이동
